refactor processMemberSources into a workflow
