+++ services/apps/premium/members_enrichment_worker/src/bin/onboarding.ts
@@ -1,16 +1,18 @@
 import { timeout } from '@crowd/common'
 import { MemberEnrichmentSource } from '@crowd/types'
 
-import { processMemberSources } from '../activities/enrichment'
 import { svc } from '../service'
+import { processMemberSources } from '../workflows/processMemberSources'
+
+export * from '@temporalio/client'
 
 // we don't need any of these to be running like if we would run this as an actual temporal worker
 // we just need pg connection, redis & service logger
-process.env['CROWD_TEMPORAL_TASKQUEUE'] = 'not-important'
+process.env['CROWD_TEMPORAL_TASKQUEUE'] = 'members-enrichment'
 svc.config.envvars = []
 svc.config.producer = { enabled: false }
 svc.config.redis = { enabled: true }
-svc.config.temporal = { enabled: false }
+svc.config.temporal = { enabled: true }
 svc.config.questdb = { enabled: false }
 svc.options.opensearch = { enabled: false }
 
@@ -88,7 +90,7 @@ setImmediate(async () => {
       processingCount++
       const startTime = Date.now()
 
-      const promise = processMemberSources(memberId, sources)
+      const promise = startProcessMemberSource(memberId, sources)
         .then((res) => {
           processingCount--
           if (res) {
@@ -127,7 +129,7 @@ setImmediate(async () => {
     await Promise.all(pagePromises)
     pagePromises = []
     // load next page
-    members = await getEnrichableMembers(pageSize, members[members.length - 1])
+    members = []
 
     svc.log.info(
       {
@@ -154,3 +156,34 @@ setImmediate(async () => {
 
   process.exit(0)
 })
+
+async function startProcessMemberSource(
+  memberId: string,
+  sources: MemberEnrichmentSource[],
+): Promise<boolean> {
+  await svc.temporal.workflow.execute(processMemberSources, {
+    taskQueue: 'members-enrichment',
+    workflowId:
+      'member-enrichment/875c38bd-2b1b-4e91-ad07-0cfbabb4c49f/' +
+      memberId +
+      '/processMemberSources',
+    workflowExecutionTimeout: '15 minutes',
+    retry: {
+      backoffCoefficient: 2,
+      maximumAttempts: 10,
+      initialInterval: 2 * 1000,
+      maximumInterval: 30 * 1000,
+    },
+    args: [
+      {
+        memberId,
+        sources,
+      },
+    ],
+    searchAttributes: {
+      TenantId: ['875c38bd-2b1b-4e91-ad07-0cfbabb4c49f'],
+    },
+  })
+
+  return true
+}
